refactor: enum 위치 변경으로 인한 import 수정
