+++ bucketback-api/src/main/java/com/programmers/bucketback/domains/bucket/application/BucketService.java
@@ -2,9 +2,9 @@ package com.programmers.bucketback.domains.bucket.application;
 
 import org.springframework.stereotype.Service;
 
-import com.programmers.bucketback.Hobby;
 import com.programmers.bucketback.common.cursor.CursorPageParameters;
 import com.programmers.bucketback.common.cursor.CursorSummary;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.common.model.ItemIdRegistry;
 import com.programmers.bucketback.domains.bucket.domain.BucketInfo;
 import com.programmers.bucketback.domains.bucket.implementation.BucketAppender;
