+++ bucketback-api/src/main/java/com/programmers/bucketback/domains/inventory/application/InventoryService.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.programmers.bucketback.Hobby;
 import com.programmers.bucketback.common.cursor.CursorPageParameters;
 import com.programmers.bucketback.common.cursor.CursorSummary;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.common.model.ItemIdRegistry;
 import com.programmers.bucketback.domains.inventory.implementation.InventoryAppender;
 import com.programmers.bucketback.domains.inventory.implementation.InventoryModifier;
