+++ bucketback-api/src/main/java/com/programmers/bucketback/domains/item/application/dto/ItemEnrollServiceRequest.java
@@ -1,6 +1,6 @@
 package com.programmers.bucketback.domains.item.application.dto;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 
 public record ItemEnrollServiceRequest(
 	Hobby hobby,
