+++ bucketback-api/src/main/java/com/programmers/bucketback/domains/vote/api/dto/request/VoteCreateRequest.java
@@ -1,6 +1,6 @@
 package com.programmers.bucketback.domains.vote.api.dto.request;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.domains.vote.application.dto.request.VoteCreateServiceRequest;
 
 import io.swagger.v3.oas.annotations.media.Schema;
