+++ bucketback-api/src/main/java/com/programmers/bucketback/domains/vote/application/dto/request/VoteCreateServiceRequest.java
@@ -1,6 +1,6 @@
 package com.programmers.bucketback.domains.vote.application.dto.request;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.domains.vote.model.VoteCreateImplRequest;
 
 import lombok.Builder;
