+++ bucketback-common/build.gradle
@@ -1,5 +1,4 @@
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
 }
 
 bootJar {
