+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/bucket/model/BucketProfile.java
@@ -2,7 +2,7 @@ package com.programmers.bucketback.domains.bucket.model;
 
 import java.util.List;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.domains.bucket.domain.Bucket;
 
 public record BucketProfile(
