+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/feed/repository/FeedRepositoryForCursor.java
@@ -2,7 +2,7 @@ package com.programmers.bucketback.domains.feed.repository;
 
 import java.util.List;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.domains.feed.model.FeedCursorSummary;
 import com.programmers.bucketback.domains.feed.model.FeedSortCondition;
