+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/inventory/model/InventoryProfile.java
@@ -2,7 +2,7 @@ package com.programmers.bucketback.domains.inventory.model;
 
 import java.util.List;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.domains.inventory.domain.Inventory;
 
 public record InventoryProfile(
