+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/vote/domain/Vote.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.domains.BaseEntity;
 import com.programmers.bucketback.domains.vote.domain.vo.Content;
