+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/vote/implementation/VoteReader.java
@@ -6,10 +6,10 @@ import java.util.Optional;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.programmers.bucketback.Hobby;
 import com.programmers.bucketback.common.cursor.CursorPageParameters;
 import com.programmers.bucketback.common.cursor.CursorSummary;
 import com.programmers.bucketback.common.cursor.CursorUtils;
+import com.programmers.bucketback.common.model.Hobby;
 import com.programmers.bucketback.domains.item.domain.Item;
 import com.programmers.bucketback.domains.item.implementation.ItemReader;
 import com.programmers.bucketback.domains.item.model.ItemInfo;
@@ -81,7 +81,8 @@ public class VoteReader {
 			throw new BusinessException(ErrorCode.UNAUTHORIZED);
 		}
 
-		if (sortCondition == VoteSortCondition.POPULARITY && statusCondition != VoteStatusCondition.COMPLETED) { // 서비스로 빼기
+		if (sortCondition == VoteSortCondition.POPULARITY
+			&& statusCondition != VoteStatusCondition.COMPLETED) { // 서비스로 빼기
 			throw new BusinessException(ErrorCode.VOTE_CANNOT_SORT);
 		}
