+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/vote/model/VoteCreateImplRequest.java
@@ -1,6 +1,6 @@
 package com.programmers.bucketback.domains.vote.model;
 
-import com.programmers.bucketback.Hobby;
+import com.programmers.bucketback.common.model.Hobby;
 
 public record VoteCreateImplRequest(
 	Hobby hobby,
